more refactoring of node code
